# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dosa Ji Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listAllOrders
      tags: [orders]
      summary: List all orders (shopkeeper/admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (shopkeeper/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── PAYMENT ──────────────────────────────────────────────────────────────────
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Activate/deactivate user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── DASHBOARD ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, shopkeeper, admin]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string

    # ── MENU ──────────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [menuItemId, name, price, quantity, imageUrl]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
    DeliveryAddress:
      type: object
      required: [name, phone, addressLine1, city, pincode]
      properties:
        name:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: ["string", "null"]
        city:
          type: string
        pincode:
          type: string
    OrderInput:
      type: object
      required: [items, deliveryAddress]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        deliveryAddress:
          $ref: "#/components/schemas/DeliveryAddress"
        couponCode:
          type: ["string", "null"]
        razorpayOrderId:
          type: ["string", "null"]
        razorpayPaymentId:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, userId, items, deliveryAddress, subtotal, deliveryCharge, gst, couponDiscount, total, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        deliveryAddress:
          $ref: "#/components/schemas/DeliveryAddress"
        subtotal:
          type: number
        deliveryCharge:
          type: number
        gst:
          type: number
        couponDiscount:
          type: number
        total:
          type: number
        razorpayOrderId:
          type: ["string", "null"]
        razorpayPaymentId:
          type: ["string", "null"]
        status:
          type: string
          enum: [Placed, Preparing, OutForDelivery, Delivered, Cancelled]
        createdAt:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Placed, Preparing, OutForDelivery, Delivered, Cancelled]

    # ── PAYMENT ───────────────────────────────────────────────────────────────
...
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [customer, shopkeeper, admin]
    UserStatusUpdate:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...