# lib/api-zod/src/generated/types/deliveryAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dosa Ji Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface DeliveryAddress {
  name: string;
  phone: string;
  addressLine1: string;
  /** @nullable */
  addressLine2?: string | null;
  city: string;
  pincode: string;
}