# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dosa Ji Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryAddress } from './deliveryAddress';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  items: OrderItem[];
  deliveryAddress: DeliveryAddress;
  subtotal: number;
  deliveryCharge: number;
  gst: number;
  couponDiscount: number;
  total: number;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
  status: OrderStatus;
  createdAt: string;
}